# docs/stoplight_docs/api-documentation-openapi.yaml
openapi: 3.0.0
servers:
  - url: 'https://injiverify.collab.mosip.net/v1/verify'
...
info:
  title: OpenID4VP Verifier API - Inji Verify
  version: 1.0.0
  contact:
    name: MOSIP
    url: 'https://mosip.io'
    email: mosip.emailnotifier@gmail.com
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.txt'
  description: |-
    API for VC verification and OpenID4VP v1.0.

    This version aligns with [OpenID4VP v1.0](https://openid.net/specs/openid-4-verifiable-presentations-1_0.html)
    and [SD-JWT VC draft-10](https://datatracker.ietf.org/doc/html/draft-ietf-oauth-sd-jwt-vc-10).

    **Breaking changes from previous versions:**
    - VP request endpoints versioned to `/v2/*` — not backward compatible
    - `presentationDefinition` / `presentationDefinitionId` replaced by `dcql_query`
    - `acceptVPWithoutHolderProof` removed — use `require_cryptographic_holder_binding` in DCQL
    - `client_id` with `did:` prefix rejected — use `decentralized_identifier:did:` prefix
    - `vp_formats` renamed to `vp_formats_supported`; `ldp_vp` renamed to `ldp_vc`
    - `presentation_submission` removed from VP submission — DCQL-only
    - `vp_token` in submission is now a JSON object keyed by DCQL `query_id`
paths:
...
  /vp-submission/dc-api:
    post:
      summary: Submit Authorization Response via Digital Credentials API (DC API)
      description: |-
        ## Description
        Used by the verifier SDK (browser) to submit a **Verifiable Presentation** returned from
        `navigator.credentials.get({ digital: … })`, per OpenID4VP `response_mode=dc_api`.

        Applies when the Authorization Request used `response_mode=dc_api`:
        - `client_id_prefix` must be `decentralized_identifier:` or `x509_san_dns:`
        - Authorization Request is **by reference** (signed JWT from `requestUri`)
        - `client_metadata` is **supported** (in that JWT)

        Correlates via JSON `requestId` (same role as `state` on `direct-post`). **No cookie is
        required on this endpoint** — the `transaction_id` cookie is only required on
        `/vp-session-results`.

        Shares the same validation pipeline as `POST /v2/vp-submission/direct-post` (structure,
        DCQL, nonce, holder binding). Does **not** return `redirect_uri` / `response_code`.

        ---

        ## Audience / Origin

        For sessions created with `response_mode=dc_api`:
        - Submission `Origin` (preferred) or `Referer` must be in the request's `expected_origins`
        - KB-JWT `aud` / LDP `proof.domain` must match the origin-bound audience:
          `origin:https://verify.example.com` (slash-free; trailing-slash form also accepted)

        ---

        ## Result retrieval

        After a successful submit (HTTP 200 empty body), the SDK calls `/vp-session-results`
        with the existing session cookie — same as `direct_post`.

      operationId: submitVpDcApi
      tags:
        - VP Submission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DcApiVpSubmissionRequest'
            examples:
              with_vp_token:
                summary: Success payload
                value:
                  requestId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  vp_token:
                    verifiable-id:
                      - type:
                          - VerifiablePresentation
              with_wallet_error:
                summary: Wallet protocol error
                value:
                  requestId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  error: invalid_request
                  error_description: optional
      responses:
        '200':
          description: VP submission accepted (empty body).
        '400':
          description: |-
            The submission failed. Body uses `errorCode` / `errorMessage`.

            Common DC API-specific cases:

            | errorCode | Meaning |
            |---|---|
            | `invalid_request` | Missing `requestId` |
            | `invalid_request` | Unable to determine verifier origin from Origin or Referer |
            | `invalid_request` | Submission Origin is not in expected_origins |
            | `invalid_request` | Audience / nonce / DCQL / structure validation failed |
            | `invalid_request` | VP already submitted for this requestId |
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_request_id:
                  summary: Missing requestId
                  value:
                    errorCode: invalid_request
                    errorMessage: requestId is required and cannot be empty.
                origin_not_allowed:
                  summary: Origin not in expected_origins
                  value:
                    errorCode: invalid_request
                    errorMessage: Submission Origin is not in the authorization request expected_origins.
...
components:
  schemas:
...
    DcApiVpSubmissionRequest:
      type: object
      required:
        - requestId
      oneOf:
        - required: [vp_token]
          not:
            anyOf:
              - required: [error]
              - required: [error_description]
        - required: [error]
          not:
            required: [vp_token]
      description: JSON body for `POST /vp-submission/dc-api`. Provide either `vp_token` or `error`.
      properties:
        requestId:
          type: string
          description: Same role as `state` on direct-post; correlates to the VP authorization request.
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        vp_token:
          type: object
          additionalProperties:
            type: array
            items: {}
          description: DCQL-keyed presentations (mutually exclusive with `error`).
        error:
          type: string
          description: Wallet protocol error code (mutually exclusive with `vp_token`).
          example: invalid_request
        error_description:
          type: string
          description: Optional wallet error description.
    # -------------------------------------------------------------------------
    # VP Result schemas
    # -------------------------------------------------------------------------
...
    ErrorResponse:
      type: object
      description: Standard error response body.
      properties:
        errorCode:
          type: string
          description: Machine-readable error code.
          example: INVALID_CLIENT_IDENTIFIER
        errorMessage:
          type: string
          description: Human-readable error description.
          example: 'clientId uses the deprecated did: prefix. Use decentralized_identifier:did: as per OpenID4VP v1.0 Section 5.9.3.'
        error:
          type: string
          description: Alternative error field (used in some endpoints per OpenID4VP spec).
        error_description:
          type: string
          description: Alternative error description field.
...